requirements.txt
django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pandas==2.2.0
networkx==3.2.1
psycopg2-binary==2.9.9